requirements.txt
mtm
opencv-python
numpy
ultralytics
pillow
Multi-Template-Matching
pandas